Change delete playlist logic

+++ Azimuth/Services/Concrete/PlaylistService.cs
@@ -122,9 +122,9 @@ namespace Azimuth.Services.Concrete
             {
                 if (AzimuthIdentity.Current != null)
                 {
-                    
-                    var userId =
-                        _userRepository.GetOne(u => u.Email.Equals(AzimuthIdentity.Current.UserCredential.Email)).Id;
+
+                    var userId = AzimuthIdentity.Current.UserCredential.Id;
+                        //_userRepository.GetOne(u => u.Email.Equals(AzimuthIdentity.Current.UserCredential.Email)).Id;
                     var currentUser = _userRepository.Get(userId);
                     var likedPlaylists = _likesRepository.Get(item => item.Liker.Id == currentUser.Id && item.IsFavorite).Select(item=>item.Playlist).ToList();
                     var playlists = _playlistRepository.Get(i => likedPlaylists.Any(j => i.Id == j.Id)).Select(playlist => Mapper.Map(playlist, new PlaylistData())).ToList();
@@ -276,13 +276,27 @@ namespace Azimuth.Services.Concrete
                     var userId = AzimuthIdentity.Current.UserCredential.Id;
                     if (userId == playlist.Creator.Id)
                     {
-                        var user = playlist.Creator;
-                        _playlistRepository.DeleteItem(playlist);
-                        _notificationService.CreateNotification(Notifications.PlaylistRemoved, user);
+                        var playlistFollowing = _likesRepository.Get(pl => pl.Playlist.Id == playlist.Id && (pl.IsFavorite || pl.IsLiked)).Count();
+                        if (playlistFollowing == 0)
+                        {
+                            _playlistRepository.DeleteItem(playlist);
+                        }
+                        else
+                        {
+                            var user = playlist.Creator;
+                            var admin =
+                                _userRepository.Get(
+                                    sysAdmin => sysAdmin.Name.FirstName.ToLower() == "admin" && sysAdmin.Name.LastName.ToLower() == "admin").FirstOrDefault();
+                            playlist.Creator = admin;
+                            _notificationService.CreateNotification(Notifications.PlaylistRemoved, user);
+                        }
                     }
                     else
                     {
-                        throw new PrivilegeNotHeldException("Only creator can delete public playlist");
+                        var favouritedPlaylist =
+                            _likesRepository.Get(record => record.Liker.Id == userId && record.IsFavorite).FirstOrDefault();
+                        if (favouritedPlaylist != null)
+                            _likesRepository.DeleteItem(favouritedPlaylist);
                     }
                 }
                 else
